+++ changelog.md
@@ -1,6 +1,6 @@
 #### 0.0.5 (2021-03-21)
 
-- preserve possible appended url fragments in converted urls (#2be45d4)
+- preserve possible appended url fragments in converted urls (#6b78f2a)
 - `pre_get_posts`: fall back to post_type `post` if empty (#000d6b5)
 - add rewrite endpoints to converted urls (#e6d4d9b)
 - add new API function `acfml_get_permalinks` (#f955315)